Start reorganizing reading module
